refactor : 불필요한 메서드, 클래스 제거 및 코드 스타일 조정
