+++ src/main/java/roomescape/domain/Password.java
@@ -1,6 +1,6 @@
 package roomescape.domain;
 
-import roomescape.exception.InvalidInputException;
+import roomescape.exception.exceptions.InvalidInputException;
 
 public class Password {
