+++ src/main/java/roomescape/domain/ReservationTime.java
@@ -1,7 +1,7 @@
 package roomescape.domain;
 
 import java.time.LocalTime;
-import roomescape.exception.InvalidInputException;
+import roomescape.exception.exceptions.InvalidInputException;
 
 public class ReservationTime {
