+++ src/main/java/roomescape/domain/UserName.java
@@ -1,6 +1,6 @@
 package roomescape.domain;
 
-import roomescape.exception.InvalidInputException;
+import roomescape.exception.exceptions.InvalidInputException;
 
 public class UserName {
