--- src/main/java/roomescape/dto/MemberCreateRequest.java
@@ -1,4 +1,0 @@
-package roomescape.dto;
-
-public record MemberCreateRequest(String name, String email, String password) {
-}
