--- src/main/java/roomescape/exception/AuthorizationException.java
@@ -1,7 +1,0 @@
-package roomescape.exception;
-
-public class AuthorizationException extends CustomException{
-    public AuthorizationException(String message) {
-        super(message);
-    }
-}
