+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -8,6 +8,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import roomescape.exception.exceptions.AuthorizationException;
+import roomescape.exception.exceptions.CustomException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
