+++ src/main/java/roomescape/exception/exceptions/AuthorizationException.java
@@ -1,0 +1,7 @@
+package roomescape.exception.exceptions;
+
+public class AuthorizationException extends CustomException {
+    public AuthorizationException(String message) {
+        super(message);
+    }
+}
