+++ src/main/java/roomescape/exception/exceptions/CustomException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exceptions;
 
 public class CustomException extends RuntimeException {
     public CustomException(String message) {
