+++ src/main/java/roomescape/exception/exceptions/ExistingEntryException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exceptions;
 
 public class ExistingEntryException extends CustomException {
     public ExistingEntryException(String message) {
