+++ src/main/java/roomescape/exception/exceptions/InvalidInputException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exceptions;
 
 public class InvalidInputException extends CustomException {
