+++ src/main/java/roomescape/exception/exceptions/NotExistingEntryException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exceptions;
 
 public class NotExistingEntryException extends CustomException {
