+++ src/main/java/roomescape/exception/exceptions/ReservingPastTimeException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exceptions;
 
 public class ReservingPastTimeException extends CustomException {
