+++ src/main/java/roomescape/service/ReservationService.java
@@ -17,9 +17,9 @@ import roomescape.dto.ReservationAdminCreateRequest;
 import roomescape.dto.ReservationCreateRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.dto.ReservationResponses;
-import roomescape.exception.InvalidInputException;
-import roomescape.exception.NotExistingEntryException;
-import roomescape.exception.ReservingPastTimeException;
+import roomescape.exception.exceptions.InvalidInputException;
+import roomescape.exception.exceptions.NotExistingEntryException;
+import roomescape.exception.exceptions.ReservingPastTimeException;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
