+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.exception.InvalidInputException;
+import roomescape.exception.exceptions.InvalidInputException;
 
 class ReservationTimeTest {
